Add dedicated /tests route, move test cards from homepage
- Create TestsPage component at /tests with all test cards
- Remove Featured Tests section from homepage (keep hero only)
- Update nav link from /#tests to /tests
- Update E2E tests to navigate to /tests for card visibility checks
